handle non-closed verbatim blocks and nil-ify empty labels
The idea behind the ``` marker is to toggle the preformatted bit on
and off, so it's not strictly speaking a "verbatim block", even if
it's easier to manage it that way.

However, don't throw generic errors like that, and handle EOF without
a closing ``` marker.  While here, if label is the empty string, use
nil.